Vim9: use :legacy for temp internal :let assignments
When changing the encoding during startup on Windows, Vim will
temporarily reset the 'isfname' and 'wildignore' options values and uses
the `:let` Vimscript command to achieve this.

However, if one is using Vim9Script initialization file, this results
in an error, because `:let` is no longer valid for Vim9Script.

Vim introduced the `:legacy` command modifier in patch 8.2.2805 so make
use of it here.

The scope can be global, since the variables are deleted after expanding
the argument list, so make it explicit as well.

Note: :unlet does not need the `:legacy` modify, as this is supported
for Vim9Script to delete the variables.

fixes: #9068