terminal: Check for valid terminal rows options
When calling :call term_start('sh', #{term_rows: 10.0}) the term_row
variable will be checked, that it is of the correct type, but later, it
will be assigned to term_tl_rows, even if it is 0. This causes later
problems when trying to allocate the memory. So make sure the actual
size is larger than zero.

fixes: #9116