formatting: use correct indenting for first line
If the buffer to be re-formatted, uses a custom indent expression,
it may happen that reformatting using e.g. `gqap` for the current
paragraph will indent the first line according to the current indent
setting (e.g. zero indent), while all the other lines will be
reformatted by using the indent expression.

This is because Vim will first join the whole paragraph into a long line
then add line breaks and reformat using the buffers current indenting
options.

This causes a nasty indent, of the first line being not indent, while
all the other ones are.

Fix this, by making sure, that even for the first line in the range,
apply the correct indenting function. That is, if an indent expression
has been set, use this, instead of the very simple get_indent() that
only indents according to the existing current indent.

Add a test to demonstrate the problem.

This fixes #9056