Reimplementation of texinfo.vim (syntax file for Texinfo).
The previous version of the Texinfo syntax file assigned
general text to the texinfoIdent syntax group and
linked to Identifier, which is inappropriate.
Also there was a problem with multiple arguments of
an @-command being handled inconsistently.

This new version of texinfo.vim does not assign general text
to a syntax group, so it is treated as Normal.
Also, multiple arguments of @-commands are colored consistently.