breakindent: correctly cache the list pattern
Do not compile the formatlistpattern every time we need to get the
indent. This just makes the breakindent function slow.

Instead, use the cached value. The indent will get re-calculated when
the line pointer changes anyhow, so it should be safe to cache the
indenting value from the last pattern match.