fix segmentation fault in tabline format parser
When parsing the statusline/tabline format string, highlight information
is stored in an array which has size 20 by default. The last item in
this array is used to indicate the end of the array, with struct members
being set to NULL/0.

When there are exactly 20 highlight items in the format string, the
array is not resized to account for the last item used to indicate the
end of the array. When copying the items to an array to be passed back
to the caller, the routine attempts to write beyond the size of the
array, often resulting in a segmentation fault.

To address this, update parser to allocate one more entry in the array
to account for the array end indicator. A test was introduced that
validates the fix when tests are run with valgrind memcheck.