fix corner case: moving on a paren from the line above after pressing `$`
Consider this text:

   xxxxx
   # (xxx)

Press `$` to jump at the end of the first line.
Press `j` to jump on the line below.
The cursor is on the closing paren.

Expected: the opening paren is highlighted.
Actual: the opening paren is NOT highlighted.

It's probably a Vim bug:
https://github.com/vim/vim/issues/5252

Asking for the syntax a second time fixes the issue.