win32: Fix slow scrolling on maximized window
When a gvim window was maximized and DirectWrite was used, scrolling
became notably slow because GDI scrolling was used unintentionally.

The edges of a maximized window are normally out of the screen.
Use the client area of the window instead of the window itself to detect
if a part of the window is off screen.