Still apply window-local directory when 'acd' fails
This fixes a regression introduced in patch 8.2.3739: window-local
directory is not applied if 'acd' fails.

It is no longer necessary to call do_autochdir() from fix_current_dir()
after patch 8.2.3920, as fix_current_dir() is now only called when
entering a window, while do_autochdir() will be called later when
entering the buffer.