don't use compound assignment expression is used inside condition
"Official" justification: https://pvs-studio.com/en/docs/warnings/v1019/

My justification: it's confusing lol. Also had an exceptionally hard
time (I'm trying not to use the word "impossible") finding information
on the priority of the calculation n-=a>=b and also what is actually
being compared with what. Like I could not for the life of me find any
rules on this. My conclusion is that it's because it's evil and must be
hidden from mankind.