Improve processing of key events in W32 GUI
This changeset adds what's essentially a replacement for
TranslateMessage that takes into account the user keyboard layout and
enables more events to be processed without having to special-case them.

The use of MapVirtualKey allows Vim to resolve some of the "virtual"
keycodes such as VK_OEM_{1..6} and friends according to the user kb
layout.

We call ToUnicode with an intentionally reduced keyboard state to avoid
a pitfall of TranslateMessage: not every combinations of Ctrl+<key>
produce a keycode (eg. Ctrl+] does not) and thus no WM_CHAR would be
emitted for them. Since we only care about the character rather than its
interaction with Ctrl or Alt let's drop those from the state matrix.

The remaining code performs what TranslateMessage would do and pump the
appropriate messages in the main loop.

Two main differences:
- Dead keys are handled in-place, no WM_DEADCHAR is emitted.
- Modifiers are handled together with the keys they're associated to
  rather than leaving them in the input queue.