Fixed foldmethod=indent fold splitting bug
The bug was introduced in patch 8.2.1560
9c2b06637b32742cac11bfd66b1a4e84583c6c2e

The bug can be reproduced by setting "foldmethod=indent" and copying line 1
and pasting it under line 2 in the following text:
line 1
        line 2
        line 3
                line 4
                line 5

"foldFind()" is used here for its side effect of setting the given fold pointer,
not for its return value. Relevant here is that the function returns null in its
final case (when the binary search fails), yet still sets the pointer. We need
that functionality to not have this bug.

In keeping with what I understand the purpose of the original patch to be, I've
kept the conditional so that the pointer arithmetic doesn't happen on a null
pointer.