win32: Handle horiziontal scroll events
Also change how the scroll events are handled: instead of directly
shifting the window we now emit the respective key event whenever the
user is scrolling gVim's window horizontally or vertically.
This allows the user to map the keys and, if needed, to block the scroll
event like it's always been possible to do on other platforms.