Fixed foldmethod=indent fold update bug
The bug is described in Github issue #3214.

The cause was that foldUpdateIEMS(), called on changed lines, would not
consider the fact that even if "top" had a defined fold level, the line
above it may not. So if you update the folding for "top", you also have
to update the folding for the line above it if the line above it has an
undefined fold level, since that line's folding relies on the line underneath.

The workarounds described in the Github issue worked because in those
cases, foldUpdateIEMS() would be called with "top" as either the blank line
(with an undefined fold level) or a line above it, instead of the line
underneath it with a defined fold level.

I didn't make a conditional to check if the line above has an undefined
fold level for whether to go one line up, because it's not worth the extra
complexity, and the similar statement on line 2240 doesn't have a conditional.