regex: \%G ignores diacrectic chars
In #8026, a way to ignore diacretic characters was requested.

Here, let's make use of the \%G atom, to ignore diacrectics when a
simple character is found.  Internally, this works by making use of
equivalence classes for each character (similar to as if you would add
[[=<char>=]] around each character.

The implementation is a bit messy, since I do not know the regexp code
very well. Also using \%G changes quite a bit how the pattern matching
works and how to count the size of the regexp beforehand, this needs to
be at the very beginning of the pattern (only after the engine
selection), it can't be done in the middle of the pattern once found,
since it changes significantly how to match ordinary items.

This is currently work in progress, so no tests are here. I just like to
get some feedback, if the approach used here seems okay or not.

Also test that \%G is at the start of the pattern, else E1515 will be
thrown.