Fixed bugs where small fold is closable after deletions
Bug 1: After changing lines, foldUpdate() is meant to set "fd_small" to "MAYBE"
on the effected folds. That code assumed that "top" was not greater than "bot",
but when foldUpdate() is called for deleted lines, "top" can be greater than
"bot".

Bug 2: The second part of the loop conditional was off by 1.