Add runtime files for bitbake
bitbake embeds Python and uses slightly different rules for indentation.
Move the GetPythonIndent function to an autoload function and allow it
to take an optional second funcref parameter. Python leaves this
argument empty so it behaves as before, but bitbake passes its own
function to affect the modified indentation rules.