po/Makefile: Avoid race over LINGUAS file
The creation of the LINGUAS file is duplicated for each desktop file
which can lead the commands to race against each other. One target might
remove it before another has been able to use it. Rework the makefile to
avoid this as the expense of leaving the file on disk.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>