Help tag generation: skip over examples
When generating the help tags file, don't parse tags inside examples.

This applies both to the implementation of the ":helptags" command, and
to the runtime/doc/doctags.c program.

As can be seen from the changes to the help tags file also included in
this commit, this removes a few bad entries from it that are not real
help tags, and which therefore cannot be jumped to.

Also add a test for this fix.