W32 GUI: try (again) to translate SPACE+Alt as well
fixes: #10753 (Vim 9.0 : cannot use AltGr+Space in gvim.exe)

Due to the fact that modifiers AltGr vs Ctrl+Alt handling was made less
sensitive in the past (see patch 8.2.4843: treating CTRL + ALT as AltGr
is not backwards compatible), special chars loop has stopped to skip
space key if it was press with AltGr. Now started to skip it again thus
translating with ToUnicode() - and bepo made happy with their '_'.