Fix color issues with GTK2 invert rectangle function
This fixes GTK2-specific part of issue #11110 by switching the "function" of
gdk_gc_new_with_values() from GDK_XOR to GDK_INVERT. It gives an inverted
appearance when selecting the text on screens like `:highlight` and `:version`
in the GTK2 build of GVim. The effect is fairly similar to the appearance to the
previous XOR'd version, but has less issues with color, especially with subpixel
anti-aliased text.

Tested with: vim --clean -g +"colo desert" +"hi"
(Also try :colo blue)