Always include FEAT_CMDWIN, alias 'tiny' build to 'small'
FEAT_CMDWIN is the only difference between the "tiny" and "small"
versions, and the difference in generated code is very small (4 to 8K,
depending on compile flags):j

	      | -O2                         | -Os
	tiny  | 1633984 (1.6M)              | 1190576 (1.2M)
	small | 1638240 (1.6M), +4256 bytes | 1198928 (1.2M), +8352 bytes

Even on very small and constrained systems this increase is negligible,
and there are quite a few #ifdefs that can be removed.

Since the 'tiny' and 'small' builds are now identical this removes the
'tiny' build altogether, but make --with-features=tiny behave like
'small' was given for compatibility so no one's build should break.