Add sound support for macOS
This finally adds +sound support for Vim on macOS. In macOS, there are
*quite* a few audio libraries to choose from. In the end, I decided to
go with AppKit which is the user-level library which we already use for
interfacing with the system clipboard. That's because the NSSound class
from it is the easiest to use and essentially plug-and-play (with the
caveat of needing a run loop as described below). Other libraries like
CoreAudio are much too powerful and overkill for Vim's use case of just
needing to play a file and be done with it, and they also involve a bit
more code to set up and use.

One caveat of using NSSound is that it's an Objective C class so I had
to do the implementation in os_macosx.m, and call that to/from sound.c.
In order for the sound callback to work, I also had to set up a run loop
that's ticked periodically in order to make sure the delegate will get
invoked whenever the sound finishes playing. The run loop is ticked the
same time as when we would have called `invoke_sound_callback()` in the
Linux / Canberra implementation.

For sound event names, I'm using the builtin NSSound's `soundNamed`
function which can locate the set of default installed sound files (e.g.
"Tink", "Sosumi", "Submarine", "Bottle", etc). I was looking up how to
get the more specific sounds (e.g. the "empty trash can" sound), and
I think NSSound doesn't know how to load them (I think Core Audio has
functions to load them but I don't think that's worth the trouble of
using Core Audio). Looking up from
https://macreports.com/how-to-change-system-sounds-empty-trash-screenshot-sent-message-etc-on-mac/,
it seems like those files can be located under
`/System/Library/Components/CoreAudio.component/Contents/SharedSupport/SystemSounds/`,
but I didn't feel like adding special code to map those files to event
names. Interested plugin authors or users can simply hard-code those
path names on macOS.