ralt_combo (issue #10826) solved, tested
Note, gvim9 seems to generally have a problem (even before this patch)
with AltGr handling, if operated via windows on-screen keyboard. Old
versions (very old gvim8) seems to be fine yet, perhaps it is a
consequence of us not calling TranslateMessage() windows API any more.

This drove me a bit crazy during my first test attempts (my RightAlt key
is remapped as Ctrl in the registry, so to test this fix i was trying to
take shortcut and used on-screen keyboard firstly and my tests failed;
with RightAlt enabled back in registry and using direct keyboard for
testing all is fine)