Add :confirm support for terminal buffers
Previously terminal windows never really worked well with `:confirm`
commands as most places still treated it like a buffer with modification
(and the subsequent "Yes" confirmation would subsequently fail as Vim
doesn't know how to "save" a terminal). The only place that really
worked was `:confirm qa`.

Fix up the code so most places (at least the ones that I could find)
that take `:confirm` also handles terminal buffers properly.

The list of commands modified:

- `check_changed()`, used for :q/:only/:enew/:next etc.
- `:bdelete`
- `:close`
- `:buffer`

(For `:buffer` and `:close` I had to manually call `free_terminal()`.
For `:buffer` it's because the fallthrough logic is to just hide the
buffer while the :confirm behavior is to kill the terminal, and for
`:close` that's because I find that otherwise it leaves a straggling
buffer that doesn't do anything, see below design question (3)).

Also, updated File.Close menu to be able to handle terminals as well.

Also added tests for both the :confirm and [!] behaviors for the above
commands.

Design questions / notes:

1. Terminal conf dialog boxes used to say yes/no/cancel, but I changed
   it to yes/no, because "no" and "cancel" used to do the same thing if
   you reason through the code. That said, I'm debating whether a
   yes/cancel makes more sense as usually the decision is to either
   "kill the terminal", or "cancel whatever operation you are doing". I
   think yes/cancel makes more sense personally, so if we want to do
   this I can update it (I just need to update all the localization po
   files as well).
2. When writing tests, I noticed that `:q!` on a terminal has a quirk
   where it would leave a straggler buffer that is empty, but not
   deleted. This seems different from the other ways a terminal gets
   killed. If you just exit the terminal normally (e.g. type "exit" in
   the shell), it will remove itself from the buffer list. I think this
   is a bug. For `:confirm close` I found a similar issue, but I worked
   around it by calling `free_terminal()`.
3. `term_start` allows you to set a `term_kill` option on a terminal.
   Right now, if you set it, and then do `:qa`, Vim will just quit
   without warning you about the terminal. However, just doing regular
   `:q` *will* warn you about the terminal. I'm not really sure which is
   the correct behavior so I kept it as it is, but it seems slightly
   inconsistent. It's not directly related to adding `:confirm` support
   anyway.