Make autoconf 2.71 work
Update configure.ac to work with autoconf 2.71; changes:

- Remove AC_HEADER_STDC; documented as "All hosted environments that are
  still of interest for portable code provide all of the headers
  specified in ISO C90".

- Remove AC_HEADER_TIME; documented as "‘time.h’ is universally present,
  and the systems on which ‘sys/time.h’ conflicted with ‘time.h’ are
  obsolete."

- Remove AC_ISC_POSIX; documented as "Sun dropped support for the
  obsolete INTERACTIVE Systems Corporation Unix on 2006-07-23.  New
  programs need not use this macro."

- AC_PROGRAM_EGREP is now AC_PROG_EGREP.

- AC_PROG_CC_C99 is now AC_PROG_CC; add a compatibility line for older
  versions of autoconf.

I ran "make autoconf" with 2.71 and 2.69, and both seem to work now.

Also, I see no more references to "configure.lineno" in current versions
of autoconf, and "When the value of $MAKE contains a path, configure
can't handle this" from todo.txt doesn't seem to be an issue any more.

Fixes #10528