Add documentation for autoload (and :color) auto-load packages
Autoload functionality now automatically searches 'packpath' to support
calling package functions from a vimrc file (#7193). However, the
documentation were not updated in that commit, so it's still a little
confusing to the user exactly how this works. Document that both in the
packages docs and autoload docs.

Also add some mentioning of :colorscheme's behavior in the autoload docs
as well as it also does something similar.