Fix off-by-one error in column calculation for popup_setpos in mouse
This fixes an off-by-one error arbitrarily introduced in the
popup_setpos calculation (in `get_fpos_of_mouse()`) code in mouse.c. I'm
not really sure why it's there as we intentionally decrement the column
by 1, but from looking at the code I cannot see why it's there at all,
considering the column is already 0-indexed.

Testing this in GTK and MacVim confirms that it fixes the issue, and the
function isn't used outside of popup_setpos.