Support `CSI {key} u` keys without modifiers
The kitty terminal emulator, as well as some others, suppors sending
keys as `CSI {key} u`, like the keys with modifiers that's already
supported, but for keys without any modifiers. The most notable example
is escape being sent as `CSI 27 u`, to be able to distinguish the key
press from the start of an escape sequence. It's also used to support
some "special" keys, like media keys and print screen.

The main reason for adding support for this is a problem with kitty and
the existing key handling in vim. By default kitty has support for keys
that are not traditionally supported in terminals, like super+<key>, and
sends those as keys with modifiers. E.g. super+q is `CSI 113 ; 9 u`. If
I press such a key, vim sets `seenModifyOtherKeys` to true and stops
processing normal control keys, so the bindings i have for those stops
working because kitty sends them as normal control keys by default.

I can set kitty to send control keys as keys with modifiers by enabling
disambiguate escape codes by sending `CSI > 1 u`. This makes the control
keys work again. However, this also changes the escape key to be sent as
`CSI 27 u`, so that stops working. By adding support for this, all keys
work as far as I know.

The problem exists by default when I'm not enabling disambiguate escape
codes though, so it would be nice if vim could enable this
automatically. I think it should be safe to send it to any terminal
emulator, but alternatively vim could query for support of the protocol
by sending `CSI ? u` and check if it gets a reply like `CSI ? flags u`.

Note that none of the escape sequences I wrote actually contains spaces,
they were just added for readability.

You can read more about the keyboard handling in kitty here:
https://sw.kovidgoyal.net/kitty/keyboard-protocol/

There's also an issue in kitty's repo with some discussion here:
https://github.com/kovidgoyal/kitty/issues/4075