Don't do anything on :!
If I run a command with :!

        :!echo asd
        asd

        Press ENTER or type command to continue

I can repeat it quickly with :!!

        :!!
        :!echo asd
        asd

        Press ENTER or type command to continue

Now I accidentally use :! instead of :!!

        :!

        Press ENTER or type command to continue

And the previous command is gone:

        :!!

        Press ENTER or type command to continue

Need to hit ENTER to clear the prompt, press :!, up, Enter to get it
back. Since this runs just "sh -c ''", which is never useful, just make
:! a no-op.

Been bugging me for years.