Update os_win32.c
fixed up a few indentation alignment issues caused by inconsistent use of tabs and spaces.  Actually, I'd slso like to fix up more in this file - in the areas that I did not touch for my main feature changes - so pending approval from maintainers - because that relates to previous code that I did not author.