Do not apply 'langmap' if merging modifiers results in another char
This is an approach to fix #11395 that makes 'langmap' behavior
consistent only for keys for which 'langmap' previously behaves
inconsistently. However, one may argue that this is now inconsistent
between different keys or when compiled with different flags. A more
consistent solution is to just avoid 'langmap' for any key that has a
modifier, but this is a bigger change in behavior.