Only set seenModifyOtherKeys when it's safe
There are certain cases where a key with modifier is seen, but keys with
the Ctrl modifier are still sent as control characters, so if
seenModifyOtherKeys is set, mappings that include the Ctrl modifier will
stop working.

There's three known cases of this:

1. The kitty terminal emulator by default, when the kitty keyboard
   protocol is not enabled by the user in the vim config, because kitty
   sends certain keys not traditionally supported by terminals as keys
   with modifiers by default and kitty doesn't support modifyOtherKeys 2.

2. The foot terminal emulator by default, because foot has
   modifyOtherKeys 1 enabled, and modifyOtherKeys 2 is not enabled
   unless done so by the user in the vim config, because vim only
   enables modifyOtherKeys 2 when TERM looks like an xterm name.

3. In a certain configuration of xterm and vim where modifyOtherKeys 1
   is enabled in xterms config, and vim is configured not to enable
   modifyOtherKeys 2. This is probably not likely to happen, but just to
   note that it is possible to encounter this in xterm too.

This patch makes sure that seenModifyOtherKeys is only set when the
control sequence to enable either kittys keyboard protocol or
modifyOtherKeys has been sent.

Fixes #9014, fixes #11403