Fix smoothscroll cursor calculations bugs when 'number' is set
The 'smoothscroll' feature currently doesn't work well when 'number' (or
signs) is set. If you scroll using `<C-E>` when the cursor is on the
first line, the cursor would end up at a wrong position, and there's
also a disconnect between the window column calculation (`wincol()`) and
the actual buffer position (`col('.')`). Commands like `g0` and `g$`
would also start to behave incorrectly.

This was a regression caused by v9.0.0746 (856c5d2bc7). That commit
actually fixed this issue for Vi compatibility mode `set cpo+=n`, but
broke the normal nocompatible case. Fixed the logic to work in both case
by correctly subtracting from `width2` instead of the entire window's
width. This makes the calculation consistent with how the "long line
wrapping" code below calculates the overflow wcol calculation.

For tests, fixed up an existing screen dump test (the old dump was
actually demonstrating the incorrect behavior, but it is quite hard to
tell IMO, because `term_dumpload` does *not* show the cursor position
and therefore it's hard to visually inspect). Also, add new tests to
confirm that the cursor position is calculated correctly and we don't
have a desync between `col('.')` and `wincol()`. Don't use screen dump
tests for these because they are testing specific logic instead of an
end-to-end tests like screen dumps, and asserts are clearer and easier
to specify this way.