Fix smoothscroll long line move cursor down out of bounds bug
With smoothscroll set, if you have a really long line that spans more
than the entire window when wrapped, if you use `j` to move cursor one
or more lines, the cursor will end up being out of bounds.

This change fixes it by properly calculating how many out-of-bounds
lines we have when scrolling down (in `scroll_cursor_bot()`). We need to
scroll that many lines in addition to the existing ones. That's because
when `smoothscroll` is on, `scrollup()` starts to only scroll by screen
lines and not buffer lines, so we need to account for how many extra
screen lines to scroll just to get past the hump of the current line, so
to speak.

If the current line is only barely long enough to span a little more
than the window size, the new cursor position will be at the bottom. If
the current line is really long, Vim will use the "center cursor" logic,
but because it doesn't really know how to handle smoothscroll right now,
it will just place the cursor at the top of teh screen instead. This is
similar to `zz` and `zb` right now as those commands are also not
smoothscroll-aware. A problem for another time to fix.

Also, add regression tests by extending the existing
`Test_smoothscroll_one_long_line()`.