Fix mouse shape test timer issue in CI
In MacVim CI, the `Test_mouse_shape_after_failed_change` tests would
fail due to test logic relying on precise timer sequencing. If any timer
gets delayed, it will cause the logic to not work. Fix this by only
starting a new timer when a previous step has completed (relying on
callbacks), instead of scheduling all timer events right off the bat.
This makes sure that we have the minimum of elapsed time between events.
Also switched to vim9script because it's easier to do a callback style
like this in it.