Reset postponed_split for failed <C-W>] commands
This prevents `postponed_split` affecting later, unrelated commands.

To reproduce, prior to this commit:

In a help file, `Ctrl-W_Ctrl-]` on a tag that doesn't exist. e.g. `:h change.txt | /CTRL-@`
(Find the "CTRL-@" and `Ctrl-W_Ctrl-]` with your cursor over it)

The tag jump will fail (this is ok/expected, since it tries to look for `CTRL-@)`)

Then, run a help command, such as `:h 'mod'`.
The help will open in a new window, not the current one