fix: avoid conflict with supercollider and scala files
Since the changes introduced in
https://github.com/vim/vim/pull/10142/commits/919ebcb8cbb5710eeccbf9750633e56572ccd471
I've seen an increasing amount of Scala users report that their Scala
scripts are being treated as supercollider files. There is one line in
the check that is problematic since it treats the following (valid
Scala) as supercollider code:

```scala
val example : String = "hello"
```

It's the ` : ` that is problematic here since looking at the notes this
can also be used for `Class : Method` in supercollider, but is also
totally valid Scala. The rest of the checks can't be confused like this
and are safe, but this one is problematic. I think it should be fine to
remove this and instead rely on the other checks to find supercollider
usage where there can't be confusion. This is anecdotal since I'm not a
supercollider user, but searching around on GitHub I'm not really even
finding this pattern for supercollider files, which again, I might be
misunderstanding their usage.