Remove 'r' and 'o' from formatoptions.
I don't think these particular options should be set from a ftplugin as
I don't believe them to be relevant to the C or Vim languages. I know
it's possible to override this with the "after" directory, but I feel
like this is a more "correct" behavior by not setting these in the first
place.