fix: change supercollider collision from \w*  to \w\+
Playing around some more with Scala / supercollider this one is also
problemantic and can be improved. For example with the current rule
using the logical `OR` on the top 25 lines causes issues. For example:

```scala
val number = 1

if (number == 2 || number = 5)
  doSomething
```

Wouldn't be recognized as a Scala file because of the `||`. Looking at
[examples](https://supercollider.github.io/examples) of supercollider it
seems that the pattern they often use for pipes is:

```supercollider
g = { | func, n = 5 |
  n.do { func = func <> func };
  func
};
```

So it would seem to make more sense to change the `\w*` to `\w+` to
ensure there is at least one between the `||` to not catch the `OR`
here.